Refactor better naming, remove hardcoded things
