+++ pom.xml
@@ -153,6 +153,8 @@ SOFTWARE.
           <systemPropertyVariables>
             <BOT_TOKEN>6018071095:AAEUVFjD3U3jNp1atDfVdJ2qxUsdkprCfMU</BOT_TOKEN>
             <BOT_NAME>blamer_dev_io_bot</BOT_NAME>
+            <HUB_HOST>localhost</HUB_HOST>
+            <HUB_PORT>7070</HUB_PORT>
           </systemPropertyVariables>
         </configuration>
       </plugin>
