+++ src/main/java/io/blamer/bot/conversation/routes/Registry.java
@@ -80,7 +80,10 @@ public class Registry implements Conversation {
         .map(auth -> new SendMessage(auth.getChat(), auth.getText()))
         .block();
     } catch (final RuntimeException ex) {
-      return new SendMessage(chat, "Internal error...");
+      return new SendMessage(
+        chat,
+        "Internal error: %s".formatted(ex.getMessage())
+      );
     }
   }
